Update run_nccl_on_idle_vms.py
Added check

+++ benchmarking/NDv4/cc-slurm-ngc/util_scripts/run_nccl_on_idle_vms.py
@@ -5,6 +5,7 @@ import subprocess
 import re
 import shutil
 from pathlib import Path
+import os
 
 # Run sinfo and get idle VMs
 cmd="sinfo"
@@ -26,7 +27,8 @@ print("VMs values: {}".format(vms_values))
 
 # Run nccl job on each VM
 dir_name = 'nccl-tests'
-shutil.rmtree(dir_name)
+if os.path.isdir(dir_name):
+    shutil.rmtree(dir_name)
 Path( dir_name ).mkdir( parents=True, exist_ok=True )
 vm_list = vms_values.split(',')
 all_vms = []
